refactor manage data dialog to its own page
