+++ src/Aspire.Dashboard/Components/Dialogs/SettingsDialog.razor
@@ -36,7 +36,7 @@
             <span class="setting-title">@Loc[nameof(Dialogs.SettingsDialogDashboardLogsAndTelemetry)]</span>
             <div>
                 <FluentButton IconStart="@(new Icons.Regular.Size16.Settings())"
-                              OnClick="LaunchManageDataAsync"
+                              OnClick="NavigateToManageData"
                               Title="@Loc[nameof(Dialogs.ManageDataManageButtonText)]"
                               aria-label="@Loc[nameof(Dialogs.ManageDataManageButtonText)]">
                     @Loc[nameof(Dialogs.ManageDataManageButtonText)]
