+++ src/Shared/DashboardUrls.cs
@@ -15,6 +15,9 @@ internal static class DashboardUrls
     public const string TracesBasePath = "traces";
     public const string LoginBasePath = "login";
     public const string HealthBasePath = "health";
+    public const string ManageDataBasePath = "managedata";
+
+    public static string ManageDataUrl() => $"/{ManageDataBasePath}";
 
     public static string ResourcesUrl(string? resource = null, string? view = null, string? hiddenTypes = null, string? hiddenStates = null, string? hiddenHealthStates = null)
     {
